format cleanup and slight re-org
